[JENKINS-58882] Add resolution for required dependency upper bounds
Why maven cannot do this on its own is beyond me. What the heck is the
point of a dependency management system if it requires the user to work
this out themselves?

The rule clearly knows about the conflict, and also assumes backwards
compatibility so why not simply *pick* the higher version? Instead I
need to declare dependencies in my POM that I *do not* care about. I do
not depend on animal-sniffer-annotations. I depend on the Truth library
that in turn depends on another library that in turn depends on
animal-sniffer. But now I need to surface dependencies two levels deep
just to satisfy this BS rule.