[JENKINS-58882] update plugin dependencies (#15)
* [JENKINS-58882] Bump base version of Jenkins 2.138

* [JENKINS-58882] Add Jenkins BOM

This specifies the versions of the dependant Jenkins plugins without us
having to do so.

* [JENKINS-58882] Increase version of Google Truth

* [JENKINS-58882] Reformat file

* [JENKINS-58882] Add resolution for required dependency upper bounds

Why maven cannot do this on its own is beyond me. What the heck is the
point of a dependency management system if it requires the user to work
this out themselves?

The rule clearly knows about the conflict, and also assumes backwards
compatibility so why not simply *pick* the higher version? Instead I
need to declare dependencies in my POM that I *do not* care about. I do
not depend on animal-sniffer-annotations. I depend on the Truth library
that in turn depends on another library that in turn depends on
animal-sniffer. But now I need to surface dependencies two levels deep
just to satisfy this BS rule.

* [JENKINS-58882] Update min version of Jenkins 2.138.4

The .1 patch version caused a the plugin not to build.

* [JENKINS-58882] Attempt to switch to Azure Container Instance

This POS Jenkins infrastructure keeps failing when applying the
recommended Jenkinsfile. So we try something else that is poorly
documented but just might work. For that, is the Jenkins way!